Moved RevokedTokenModel model class out of model.py
