+++ iot_api/user_api/model.py
@@ -273,22 +273,6 @@ class UserToUserRole(db.Model):
         return cls.query.filter_by(user_id=user_id, user_role_id=user_role_id).first()
 
 
-class RevokedTokenModel(db.Model):
-    __tablename__ = "revoked_tokens"
-
-    id = db.Column(db.BigInteger, primary_key=True)
-    jti = db.Column(db.String(120))
-
-    def add(self):
-        db.session.add(self)
-        db.session.commit()
-
-    @classmethod
-    def is_jti_blacklisted(cls, jti):
-        query = cls.query.filter_by(jti=jti).first()
-        return bool(query)
-
-
 class AccountActivation(db.Model):
     __tablename__ = "account_activation"
