+++ iot_api/user_api/models/__init__.py
@@ -22,3 +22,4 @@ from iot_api.user_api.models.ChangeEmailRequests import ChangeEmailRequests
 from iot_api.user_api.models.GlobalData import GlobalData
 from iot_api.user_api.models.LoginAttempts import LoginAttempts
 from iot_api.user_api.models.PasswordReset import PasswordReset
+from iot_api.user_api.models.RevokedTokenModel import RevokedTokenModel
