+++ iot_api/user_api/resources/endpoints.py
@@ -27,7 +27,7 @@ from iot_api import config
 from iot_api.user_api.enums import RoleTypes
 from iot_api.user_api.events.data_collector_events import emit_data_collector_event
 from iot_api.user_api.events.policy_events import emit_policy_event
-from iot_api.user_api.model import User, Organization, Device, RevokedTokenModel, AccountActivation, StatsCounters, \
+from iot_api.user_api.model import User, Organization, Device, AccountActivation, StatsCounters, \
     UserRole, UserToUserRole, Alert, AlertType, Packet, \
     SendMailAttempts, \
     get_user_collector_ids
@@ -45,6 +45,7 @@ from iot_api.user_api.models.NotificationPreferences import NotificationPreferen
 from iot_api.user_api.models.PasswordReset import PasswordReset
 from iot_api.user_api.models.Policy import Policy
 from iot_api.user_api.model import Quarantine
+from iot_api.user_api.models.RevokedTokenModel import RevokedTokenModel
 from iot_api.user_api.singletonURL import singletonURL
 
 from urllib.parse import urlencode
